[ADD] delete popup traduction

+++ src/app/company/companies/companies.component.ts
@@ -4,6 +4,7 @@ import {CompanyService} from '../company.service';
 import {MatPaginator, PageEvent} from '@angular/material';
 import {Page} from '../../page.model';
 import {animate, keyframes, style, transition, trigger} from '@angular/animations';
+import {TranslateService} from '@ngx-translate/core';
 
 @Component({
   selector: 'app-companies',
@@ -38,6 +39,7 @@ export class CompaniesComponent implements OnInit {
 
 
   companies: Page<Company>;
+  //translate: TranslateService;
   length;
   pageSize = 10;
   pageSizeOptions = [4, 10, 20, 100];
@@ -46,7 +48,7 @@ export class CompaniesComponent implements OnInit {
   @ViewChild(MatPaginator) paginator: MatPaginator;
   checked = false;
 
-  constructor(private companyService: CompanyService) {
+  constructor(private companyService: CompanyService, private  translate: TranslateService) {
   }
 
   ngOnInit() {
@@ -74,7 +76,9 @@ export class CompaniesComponent implements OnInit {
 
   deleteCompany(company: Company) {
     console.log(company);
-    this.companyService.deleteCompany(company.id).subscribe();
-    this.companies.results.splice(this.companies.results.indexOf(company), 1 );
+    if(confirm( this.translate.instant('POPUP.ON_DELETE') + company.name + '?')) {
+      this.companyService.deleteCompany(company.id).subscribe();
+      this.companies.results.splice(this.companies.results.indexOf(company), 1 );
+    }
   }
 }
